Move runMLXWinotunerTests into mlxtests.cpp
Completes the MLX test-code extraction: the tuner test entry point
now lives alongside runMLXWinogradTests in the dedicated test
translation unit, leaving mlxwinotuner.cpp focused on tuner logic.
