+++ cpp/neuralnet/mlxwinotuner.cpp
@@ -20,7 +20,6 @@
 #include "../core/global.h"
 #include "../core/logger.h"
 #include "../core/makedir.h"
-#include "../core/test.h"
 #include "../dataio/homedata.h"
 
 #include "mlx/mlx.h"
@@ -1065,702 +1064,6 @@ std::string MLXWinogradTuner::formatConv3x3Distribution(const ModelDesc& modelDe
   int total = 0;
   for(const auto& kv : outVec) total += kv.second;  // total = #3x3 convs
   return formatConv3x3DistributionLine(total, inMap, outMap);
-}
-
-void runMLXWinotunerTests() {
-  cout << "Running MLX Winograd tuner tests" << endl;
-
-  {
-    // Conv-3x3 distribution formatter — pure-function test. Verifies the
-    // log-line format directly without any descriptor walk or GPU work.
-    // Order convention: pairs sorted descending by invocation
-    // count, ties broken by channel count descending.
-
-    // Case A: two distinct shapes, each appearing once. Tie on count, so
-    // tie-break by channel count descending: 64 before 32.
-    {
-      std::map<int,int> inputC  = {{32, 1}, {64, 1}};
-      std::map<int,int> outputC = {{32, 1}, {64, 1}};
-      std::string line = MLXWinogradTuner::formatConv3x3DistributionLine(2, inputC, outputC);
-      testAssert(line.find("MLX tuner conv3x3 distribution:") != std::string::npos);
-      testAssert(line.find("total=2") != std::string::npos);
-      testAssert(line.find("input_c=64:1,32:1") != std::string::npos);
-      testAssert(line.find("output_c=64:1,32:1") != std::string::npos);
-    }
-
-    // Case B: asymmetric counts. 384 appears 36 times, 192 once. Sort by
-    // count descending, so 384 first regardless of channel-count order.
-    {
-      std::map<int,int> inputC  = {{384, 36}, {192, 1}};
-      std::map<int,int> outputC = {{384, 37}};
-      std::string line = MLXWinogradTuner::formatConv3x3DistributionLine(37, inputC, outputC);
-      testAssert(line.find("total=37") != std::string::npos);
-      testAssert(line.find("input_c=384:36,192:1") != std::string::npos);
-      testAssert(line.find("output_c=384:37") != std::string::npos);
-    }
-
-    // Case C: empty model — no 3x3 convs. Error handling: print the
-    // line with explicit "{}" markers; don't suppress.
-    {
-      std::map<int,int> empty;
-      std::string line = MLXWinogradTuner::formatConv3x3DistributionLine(0, empty, empty);
-      testAssert(line.find("total=0") != std::string::npos);
-      testAssert(line.find("input_c={}") != std::string::npos);
-      testAssert(line.find("output_c={}") != std::string::npos);
-    }
-    std::cout << "  conv3x3 distribution formatter OK" << std::endl;
-  }
-
-  {
-    // planShapeRotation — pure-function tests. Verifies the selection rule
-    // (top-3, 3% threshold, 3-rep floor, proportional remainder) directly
-    // without any GPU work.
-
-    // Case A: single shape — entire budget on that shape, weight = 1.0.
-    {
-      auto plan = MLXWinogradTuner::planShapeRotationForTesting({{192, 72}});
-      testAssert(plan.size() == 1);
-      testAssert(plan[0].channels == 192);
-      testAssert(plan[0].measureReps == 19);
-      testAssert(std::abs(plan[0].weight - 1.0) < 1e-9);
-    }
-
-    // Case B: two shapes both above threshold (b18c384nbt-like, after the
-    // 22:1 entry has already been dropped by threshold). Expected:
-    // work = 192*72, 128*5 = 13824, 640; weights 0.956, 0.044;
-    // round(0.956*19)=18, round(0.044*19)=1; floor bumps 1->3; dominant 18-2=16.
-    {
-      auto plan = MLXWinogradTuner::planShapeRotationForTesting({{192, 72}, {128, 5}});
-      testAssert(plan.size() == 2);
-      testAssert(plan[0].channels == 192);
-      testAssert(plan[1].channels == 128);
-      testAssert(plan[0].measureReps == 16);
-      testAssert(plan[1].measureReps == 3);
-      testAssert(plan[0].measureReps + plan[1].measureReps == 19);
-      testAssert(std::abs(plan[0].weight + plan[1].weight - 1.0) < 1e-9);
-      testAssert(plan[0].weight > plan[1].weight);
-    }
-
-    // Case C: minor shape below 3% threshold — dropped entirely, dominant
-    // absorbs all 19 reps. Histogram: 192:72 (work 13824, 95.5%), 22:1 (work 22, 0.15%).
-    {
-      auto plan = MLXWinogradTuner::planShapeRotationForTesting({{192, 72}, {22, 1}});
-      testAssert(plan.size() == 1);
-      testAssert(plan[0].channels == 192);
-      testAssert(plan[0].measureReps == 19);
-      testAssert(std::abs(plan[0].weight - 1.0) < 1e-9);
-    }
-
-    // Case D: four shapes — top-3 cut drops the 4th, then threshold drops
-    // one more. Input: 384:60, 192:8, 128:5, 64:5. After top-3: drop 64:5.
-    // work remaining = 23040, 1536, 640; total 25216; 128's share = 2.54% < 3%
-    // -> drop 128. Final: 384 (93.75%) + 192 (6.25%). reps: round(0.9375*19)=18,
-    // round(0.0625*19)=1; floor bumps 1->3; dominant 18-2=16.
-    {
-      auto plan = MLXWinogradTuner::planShapeRotationForTesting(
-          {{384, 60}, {192, 8}, {128, 5}, {64, 5}});
-      testAssert(plan.size() == 2);
-      testAssert(plan[0].channels == 384);
-      testAssert(plan[1].channels == 192);
-      testAssert(plan[0].measureReps == 16);
-      testAssert(plan[1].measureReps == 3);
-    }
-
-    // Case E: three shapes all above threshold. Input: 200:10, 100:10, 50:10.
-    // work = 2000, 1000, 500; total 3500; shares 57.1%, 28.6%, 14.3% (all >3%).
-    // reps: round(0.571*19)=11, round(0.286*19)=5, round(0.143*19)=3.
-    // Sum = 19 exactly (no rounding repair needed). All >= floor of 3.
-    {
-      auto plan = MLXWinogradTuner::planShapeRotationForTesting(
-          {{200, 10}, {100, 10}, {50, 10}});
-      testAssert(plan.size() == 3);
-      testAssert(plan[0].channels == 200);
-      testAssert(plan[1].channels == 100);
-      testAssert(plan[2].channels == 50);
-      int total = plan[0].measureReps + plan[1].measureReps + plan[2].measureReps;
-      testAssert(total == 19);
-      testAssert(plan[2].measureReps >= 3);
-      testAssert(plan[0].measureReps >= plan[1].measureReps);
-      testAssert(plan[1].measureReps >= plan[2].measureReps);
-    }
-
-    // Case F: 2 shapes with equal work and complementary 0.5 shares —
-    // exercises the rounding-repair branch. Input: 200:1, 100:2 (work
-    // 200, 200; tied; tie-break by larger C → plan[0]=C=200). Each
-    // share is 0.5; lround(0.5*19) = lround(9.5) = 10 each (lround
-    // rounds halves away from zero); pre-repair sum = 20; repair:
-    // dominant absorbs delta = 19 - 20 = -1; final (9, 10). Both
-    // measureReps stay ≥ kRepFloor=3 so floor-bump is a no-op.
-    {
-      auto plan = MLXWinogradTuner::planShapeRotationForTesting(
-          {{200, 1}, {100, 2}});
-      testAssert(plan.size() == 2);
-      testAssert(plan[0].channels == 200);
-      testAssert(plan[1].channels == 100);
-      testAssert(plan[0].measureReps + plan[1].measureReps == 19);
-      testAssert(plan[0].measureReps == 9);
-      testAssert(plan[1].measureReps == 10);
-      testAssert(plan[0].measureReps >= 3);
-      testAssert(plan[1].measureReps >= 3);
-    }
-
-    std::cout << "  planShapeRotation OK" << std::endl;
-  }
-
-  {
-    // buildConv3x3HistogramsFromConvs — pure-function test on the conv
-    // filter+histogram. Constructs ConvLayerDesc instances directly
-    // (default-constructible per desc.h:25). ConvLayerDesc has a deleted
-    // copy ctor (desc.h:29), so we build the descriptors in a deque
-    // (stable addresses, no copies on growth) and pass pointers to the
-    // helper. Does not touch ModelDesc.
-
-    auto initConv = [](ConvLayerDesc& c, int kY, int kX, int inC, int outC) {
-      c.convYSize  = kY;
-      c.convXSize  = kX;
-      c.inChannels = inC;
-      c.outChannels = outC;
-    };
-
-    // Four layers: only the two 3x3 layers should contribute.
-    std::deque<ConvLayerDesc> storage;
-    std::vector<const ConvLayerDesc*> convs;
-    storage.emplace_back(); initConv(storage.back(), 1, 1, 10, 10); convs.push_back(&storage.back());  // 1x1 — filtered
-    storage.emplace_back(); initConv(storage.back(), 3, 3, 20, 30); convs.push_back(&storage.back());  // input_c[20]++, output_c[30]++
-    storage.emplace_back(); initConv(storage.back(), 3, 3, 30, 30); convs.push_back(&storage.back());  // input_c[30]++, output_c[30]++
-    storage.emplace_back(); initConv(storage.back(), 5, 5, 40, 40); convs.push_back(&storage.back());  // 5x5 — filtered
-
-    auto [inHist, outHist] =
-        MLXWinogradTuner::buildConv3x3HistogramsFromConvsForTesting(convs);
-
-    // Convert to maps for order-independent comparison.
-    std::map<int,int> inMap(inHist.begin(), inHist.end());
-    std::map<int,int> outMap(outHist.begin(), outHist.end());
-
-    testAssert(inMap.size() == 2);
-    testAssert(inMap[20] == 1);
-    testAssert(inMap[30] == 1);
-    testAssert(inMap.count(10) == 0);  // 1x1 didn't leak through
-    testAssert(inMap.count(40) == 0);  // 5x5 didn't leak through
-
-    testAssert(outMap.size() == 1);
-    testAssert(outMap[30] == 2);
-    testAssert(outMap.count(10) == 0);
-    testAssert(outMap.count(40) == 0);
-
-    // Asymmetric 3x3 (e.g. 3x1) must also be filtered — the kernel is
-    // strictly square-3.
-    std::deque<ConvLayerDesc> asymStorage;
-    std::vector<const ConvLayerDesc*> asym;
-    asymStorage.emplace_back(); initConv(asymStorage.back(), 3, 1, 16, 16); asym.push_back(&asymStorage.back());
-    asymStorage.emplace_back(); initConv(asymStorage.back(), 1, 3, 16, 16); asym.push_back(&asymStorage.back());
-    asymStorage.emplace_back(); initConv(asymStorage.back(), 3, 3, 16, 16); asym.push_back(&asymStorage.back());
-    auto [inA, outA] =
-        MLXWinogradTuner::buildConv3x3HistogramsFromConvsForTesting(asym);
-    testAssert(inA.size() == 1 && inA[0].first == 16 && inA[0].second == 1);
-    testAssert(outA.size() == 1 && outA[0].first == 16 && outA[0].second == 1);
-
-    // Empty input → empty histograms (no assert; this is just the pure
-    // core. The mlxbackend.cpp call site asserts non-empty after a real
-    // model walk; mlxbackend.cpp pre-computes the histogram at model
-    // load and stores it on ModelInfoForTuning so the tuner does not
-    // re-walk the descriptor).
-    std::vector<const ConvLayerDesc*> empty;
-    auto [inE, outE] =
-        MLXWinogradTuner::buildConv3x3HistogramsFromConvsForTesting(empty);
-    testAssert(inE.empty());
-    testAssert(outE.empty());
-
-    std::cout << "  buildConv3x3HistogramsFromConvs OK" << std::endl;
-  }
-
-  // ---- v3 round-trip: tg0/tg1/wpt/vw/gridOrder (input), tg0/tg1/wpt (output) ----
-  {
-    // v3 roundtrip: write -> load -> compare all 8 fields. Two
-    // cases for input gridOrder: Cfast and Tfast. (Tfast forces vw=1 per
-    // isValid invariant.)
-    using namespace MLXWinograd;
-    for(auto inGo : {GridOrder::Cfast, GridOrder::Tfast}) {
-      MLXWinogradTuneParams p;
-      p.inputTransform.tg0 = 32;
-      p.inputTransform.tg1 = 1;
-      p.inputTransform.wpt = 2;
-      p.inputTransform.vw  = (inGo == GridOrder::Cfast) ? 2 : 1;
-      p.inputTransform.gridOrder = inGo;
-      p.outputUntransform.tg0 = 32;
-      p.outputUntransform.tg1 = 8;
-      p.outputUntransform.wpt = 1;
-      testAssert(p.isValid());
-
-      std::string tmpFile = "/tmp/sp5_v3_roundtrip_" + std::to_string((int)inGo) + ".txt";
-      MLXWinogradTuneParams::save(tmpFile, p);
-      MLXWinogradTuneParams q = MLXWinogradTuneParams::load(tmpFile);
-      testAssert(q.inputTransform.tg0 == p.inputTransform.tg0);
-      testAssert(q.inputTransform.tg1 == p.inputTransform.tg1);
-      testAssert(q.inputTransform.wpt == p.inputTransform.wpt);
-      testAssert(q.inputTransform.vw  == p.inputTransform.vw);
-      testAssert(q.inputTransform.gridOrder == p.inputTransform.gridOrder);
-      testAssert(q.outputUntransform.tg0 == p.outputUntransform.tg0);
-      testAssert(q.outputUntransform.tg1 == p.outputUntransform.tg1);
-      testAssert(q.outputUntransform.wpt == p.outputUntransform.wpt);
-      testAssert(q.isValid());
-      std::remove(tmpFile.c_str());
-    }
-    cout << "  v3 roundtrip (Cfast + Tfast) OK" << endl;
-  }
-
-  // dtype-aware cache filenames must coexist in the same directory
-  // without collision. Verify defaultFileName gains a _fp16/_fp32 suffix.
-  {
-    std::string nameF32 = MLXWinogradTuner::defaultFileName(
-      "AppleSilicon", 19, 19, 384, 13, /*useFP16=*/false);
-    std::string nameF16 = MLXWinogradTuner::defaultFileName(
-      "AppleSilicon", 19, 19, 384, 13, /*useFP16=*/true);
-    testAssert(nameF32 != nameF16);
-    testAssert(nameF32.find("_fp32") != std::string::npos);
-    testAssert(nameF16.find("_fp16") != std::string::npos);
-    testAssert(nameF32.size() >= 4 && nameF32.substr(nameF32.size()-4) == ".txt");
-    testAssert(nameF16.size() >= 4 && nameF16.substr(nameF16.size()-4) == ".txt");
-    cout << "  defaultFileName dtype suffix OK: "
-         << nameF32 << " vs " << nameF16 << endl;
-  }
-
-  // ---- Corrupt-version rejection ----
-  {
-    std::string tmp = "/tmp/katago_mlx_winotuner_badversion.txt";
-    {
-      std::ofstream f(tmp);
-      f << "VERSION=999\n#inputTransform\ntg0=32 tg1=1\n#outputUntransform\ntg0=32 tg1=1\n";
-    }
-    bool threw = false;
-    try { (void)MLXWinogradTuneParams::load(tmp); }
-    catch(const IOError&) { threw = true; }
-    testAssert(threw);
-  }
-
-  // ---- v3 isValid invariants ----
-  {
-    // v3 isValid invariants.
-    using namespace MLXWinograd;
-    auto basePass = [&]() {
-      MLXWinogradTuneParams p;
-      p.inputTransform = {32, 1, 1, 2, GridOrder::Cfast};
-      p.outputUntransform = {32, 2, 1};
-      return p;
-    };
-
-    // Baseline passes.
-    testAssert(basePass().isValid());
-
-    // tg0 <= 0 fails.
-    { auto p = basePass(); p.inputTransform.tg0 = 0;  testAssert(!p.isValid()); }
-    { auto p = basePass(); p.outputUntransform.tg0 = -1; testAssert(!p.isValid()); }
-
-    // tg0 * tg1 > 1024 fails.
-    { auto p = basePass(); p.inputTransform.tg0 = 64; p.inputTransform.tg1 = 32;
-      testAssert(!p.isValid()); }
-
-    // wpt < 1 fails.
-    { auto p = basePass(); p.inputTransform.wpt = 0;  testAssert(!p.isValid()); }
-    { auto p = basePass(); p.outputUntransform.wpt = 0; testAssert(!p.isValid()); }
-
-    // vw < 1 fails on input.
-    { auto p = basePass(); p.inputTransform.vw = 0;   testAssert(!p.isValid()); }
-
-    // Tfast on input forces vw=1.
-    { auto p = basePass();
-      p.inputTransform.gridOrder = GridOrder::Tfast;
-      p.inputTransform.vw = 2;
-      testAssert(!p.isValid()); }
-    { auto p = basePass();
-      p.inputTransform.gridOrder = GridOrder::Tfast;
-      p.inputTransform.vw = 1;
-      testAssert(p.isValid()); }
-
-    cout << "  v3 isValid invariants OK" << endl;
-  }
-
-  // Candidate enumeration with validity filtering.
-  {
-    using namespace MLXWinograd;
-    // Cfast, C=64 (divisible by all vw): full Cartesian product over all axes
-    // minus tg0*tg1>1024.
-    auto cands = MLXWinogradTuner::buildInputCandidatesForTesting(
-        /*full*/true, /*C*/64, /*Ntiles*/200, GridOrder::Cfast);
-
-    // Sanity: returns hundreds of valid configs.
-    testAssert(cands.size() > 100);
-    testAssert(cands.size() < 5000);   // bounded by validity filter
-
-    // All candidates satisfy tg0*tg1 <= 1024.
-    for(const auto& c : cands)
-      testAssert(c.tg0 * c.tg1 <= 1024);
-
-    // C=66 with vw>1: should filter out vw=2 (66%2=0 — VW=2 allowed)
-    // and vw=4 (66%4=2 != 0 — VW=4 should NOT appear in candidates).
-    auto cands_C66 = MLXWinogradTuner::buildInputCandidatesForTesting(
-        true, /*C*/66, /*Ntiles*/200, GridOrder::Cfast);
-    for(const auto& c : cands_C66) {
-      if(c.vw == 4)
-        testAssert(false);  // vw=4 candidate should have been filtered out for C=66
-    }
-
-    // Tfast: vw must be 1 (kernel static_assert). All Tfast candidates have vw=1.
-    auto cands_Tfast = MLXWinogradTuner::buildInputCandidatesForTesting(
-        true, 64, 200, GridOrder::Tfast);
-    for(const auto& c : cands_Tfast) {
-      testAssert(c.vw == 1);
-      testAssert(c.gridOrder == GridOrder::Tfast);
-    }
-
-    // Output side: same shape of assertions. (gridOrder is not a parameter
-    // of buildOutputCandidatesForTesting — output is Cfast-only.)
-    auto out_cands = MLXWinogradTuner::buildOutputCandidatesForTesting(
-        true, /*outC*/64, /*Ntiles*/200);
-    testAssert(out_cands.size() > 100);
-    for(const auto& c : out_cands)
-      testAssert(c.tg0 * c.tg1 <= 1024);
-
-    std::cout << "  MLX Winograd candidate enumeration validity passed ("
-              << cands.size() << " input / " << out_cands.size() << " output candidates C=64)"
-              << std::endl;
-  }
-
-  // ---- Measurement primitives return finite positive times ----
-  // We can't call the static helpers from the test, so we use the public
-  // surface: loadOrAutoTune with reTune=true runs the search and we verify
-  // that the public schema struct works with valid configs. The measurement
-  // primitive itself is exercised by the search-works test below.
-
-  {
-    // Gated flat-sweep convergence test.
-    // Runs the production flat sweep on a small synthetic problem and asserts
-    // that the winner is isValid and that its timing is no worse than the
-    // baked default (tg0=32, tg1=1, wpt=1, vw=1, Cfast).
-    const char* gate = std::getenv("KATAGO_MLX_WINOTUNER_RUN_SWEEP_TEST");
-    if(gate != nullptr && std::string(gate) == "1") {
-      MLXWinogradTuner::ModelInfoForTuning mi;
-      mi.trunkNumChannels    = 64;
-      mi.modelVersion        = 11;
-      // Synthetic single-shape histogram for the toy C=64 test model.
-      mi.conv3x3InputHistogram  = {{64, 1}};
-      mi.conv3x3OutputHistogram = {{64, 1}};
-
-      // loadOrAutoTune rewrites an empty tunerFile to a default cache path,
-      // so use an explicit temp path and remove it after to avoid touching
-      // the user's cache directory.
-      std::string tmpTunerFile = "/tmp/sp5_task10_sweep_cache.txt";
-      std::remove(tmpTunerFile.c_str());
-
-      MLXWinogradTuneParams tuned = MLXWinogradTuner::loadOrAutoTune(
-          /*tunerFile=*/tmpTunerFile,
-          /*homeDataDirOverride=*/"",
-          /*gpuName=*/"AppleSilicon",
-          /*nnXLen=*/19, /*nnYLen=*/19, /*batchSize=*/1,
-          mi,
-          /*logger=*/nullptr,
-          /*full=*/false,
-          /*reTune=*/true,
-          /*useFP16=*/true);
-      testAssert(tuned.isValid());
-
-      // Score the baked default and the tuned winner via scoreInputTransform.
-      // tuned.time <= baked.time (within noise).
-      MLXWinograd::InputTransform baked{};
-      baked.tg0 = 32; baked.tg1 = 1; baked.wpt = 1; baked.vw = 1;
-      baked.gridOrder = MLXWinograd::GridOrder::Cfast;
-      auto bestOf5 = [&](const MLXWinograd::InputTransform& cfg) -> double {
-        double best = std::numeric_limits<double>::infinity();
-        for(int rep = 0; rep < 5; rep++) {
-          double t = MLXWinogradTuner::scoreInputTransformForTesting(
-              cfg, 1, 19, 19, mi, true);
-          if(t < best) best = t;
-        }
-        return best;
-      };
-      double bakedMs = bestOf5(baked);
-      double tunedMs = bestOf5(tuned.inputTransform);
-      // Allow 10% noise budget.
-      testAssert(tunedMs <= bakedMs * 1.10);
-      std::cout << "  flat-sweep convergence (gated) OK"
-                << " bakedMs=" << bakedMs
-                << " tunedMs=" << tunedMs << std::endl;
-
-      std::remove(tmpTunerFile.c_str());
-    }
-  }
-
-  {
-    // Baseline anchor — Test 1: log-format gated check (input stage).
-    // Asserts that flatSweepInput's log line carries the new baseline_ms and
-    // delta_pct fields with the documented format. Gated because the synthetic
-    // sweep takes a few seconds; opt in with the env var below.
-    const char* gate = std::getenv("KATAGO_MLX_WINOTUNER_RUN_LOG_FORMAT_TEST");
-    if(gate != nullptr && std::string(gate) == "1") {
-      MLXWinogradTuner::ModelInfoForTuning mi;
-      mi.trunkNumChannels    = 64;
-      mi.modelVersion        = 11;
-      // Synthetic single-shape histogram for the toy C=64 test model.
-      mi.conv3x3InputHistogram  = {{64, 1}};
-      mi.conv3x3OutputHistogram = {{64, 1}};
-
-      std::string tmpTunerFile = "/tmp/baseline_anchor_log_format.txt";
-      std::remove(tmpTunerFile.c_str());
-
-      std::ostringstream captured;
-      Logger logger(nullptr, /*logToStdoutDefault=*/false,
-                    /*logToStderrDefault=*/false, /*logTimeDefault=*/false,
-                    /*logConfigContents=*/false);
-      logger.addOStream(captured);
-
-      (void)MLXWinogradTuner::loadOrAutoTune(
-          /*tunerFile=*/tmpTunerFile,
-          /*homeDataDirOverride=*/"",
-          /*gpuName=*/"AppleSilicon",
-          /*nnXLen=*/19, /*nnYLen=*/19, /*batchSize=*/1,
-          mi,
-          /*logger=*/&logger,
-          /*full=*/false,
-          /*reTune=*/true,
-          /*useFP16=*/true);
-
-      const std::string log = captured.str();
-      // Logger::writeLocked prefixes each line with ": " when logTime=false, so
-      // `log` reads ": MLX tuner ...". std::regex_search is anchor-free so the
-      // ": " prefix is transparent; only the substring match matters here.
-      // The regex matches the non-degenerate path only (best != nullopt). The
-      // best=none / delta_pct=nan branch is unreachable for the synthetic 19x19
-      // C=64 problem this test runs against (hundreds of valid candidates).
-      // Updated for shape diagnostic: regex now requires the per-shape
-      // median fields appended by flatSweepInput.
-      std::regex inputRe(
-          R"(MLX tuner flatSweepInput: considered=[0-9]+ best=tg0=[0-9]+ tg1=[0-9]+ wpt=[0-9]+ vw=[0-9]+ gridOrder=[01] time_ms=[0-9]+\.[0-9]+ baseline_ms=[0-9]+\.[0-9]+ delta_pct=[-+][0-9]+\.[0-9]+ shape_ms=c[0-9]+:[0-9]+\.[0-9]+(?:,c[0-9]+:[0-9]+\.[0-9]+)*)");
-      testAssert(std::regex_search(log, inputRe));
-      std::cout << "  flatSweepInput log-format (gated) OK" << std::endl;
-
-      std::regex outputRe(
-          R"(MLX tuner flatSweepOutput: considered=[0-9]+ best=tg0=[0-9]+ tg1=[0-9]+ wpt=[0-9]+ time_ms=[0-9]+\.[0-9]+ baseline_ms=[0-9]+\.[0-9]+ delta_pct=[-+][0-9]+\.[0-9]+ shape_ms=c[0-9]+:[0-9]+\.[0-9]+(?:,c[0-9]+:[0-9]+\.[0-9]+)*)");
-      testAssert(std::regex_search(log, outputRe));
-      std::cout << "  flatSweepOutput log-format (gated) OK" << std::endl;
-
-      std::remove(tmpTunerFile.c_str());
-    }
-  }
-
-  {
-    // Baseline anchor — Test 2: baseline-consistency gated check.
-    // Asserts that the baseline_ms value printed by flatSweepInput
-    // matches an independent re-score of the default-constructed
-    // InputTransform within a 25% relative-error budget.
-    //
-    // parsedBaseline is a single 20-rep weighted mean (one call into
-    // scoreInputTransform). minOf3 is the min of three such weighted
-    // means — systematically biased slightly low relative to a single
-    // mean due to selection bias (~5-10% on this hardware), on top of
-    // the ~10% per-sample noise floor. The 25% budget covers both.
-    //
-    // Reuses the KATAGO_MLX_WINOTUNER_RUN_SWEEP_TEST gate so users who
-    // opt into the sweep-convergence cost also get this check. Note
-    // this runs an INDEPENDENT loadOrAutoTune sweep — total cost when
-    // the gate is set is roughly 2x the cost of a single sweep.
-    //
-    // Coverage scope: input stage only. flatSweepOutput's baseline_ms
-    // is format-checked by Test 1 but not consistency-checked here.
-    // The output kernel uses a different scoring function and default
-    // struct (OutputUntransform{}); a symmetric check is deferred.
-    const char* gate = std::getenv("KATAGO_MLX_WINOTUNER_RUN_SWEEP_TEST");
-    if(gate != nullptr && std::string(gate) == "1") {
-      MLXWinogradTuner::ModelInfoForTuning mi;
-      mi.trunkNumChannels    = 64;
-      mi.modelVersion        = 11;
-      // Synthetic single-shape histogram for the toy C=64 test model.
-      mi.conv3x3InputHistogram  = {{64, 1}};
-      mi.conv3x3OutputHistogram = {{64, 1}};
-
-      std::string tmpTunerFile = "/tmp/baseline_anchor_consistency.txt";
-      std::remove(tmpTunerFile.c_str());
-
-      std::ostringstream captured;
-      Logger logger(nullptr, /*logToStdoutDefault=*/false,
-                    /*logToStderrDefault=*/false, /*logTimeDefault=*/false,
-                    /*logConfigContents=*/false);
-      logger.addOStream(captured);
-
-      (void)MLXWinogradTuner::loadOrAutoTune(
-          /*tunerFile=*/tmpTunerFile,
-          /*homeDataDirOverride=*/"",
-          /*gpuName=*/"AppleSilicon",
-          /*nnXLen=*/19, /*nnYLen=*/19, /*batchSize=*/1,
-          mi,
-          /*logger=*/&logger,
-          /*full=*/false,
-          /*reTune=*/true,
-          /*useFP16=*/true);
-
-      const std::string log = captured.str();
-      std::smatch m;
-      std::regex baselineRe(R"(flatSweepInput:[^\n]*baseline_ms=([0-9]+\.[0-9]+))");
-      testAssert(std::regex_search(log, m, baselineRe));
-      const double parsedBaseline = std::stod(m[1].str());
-
-      double minOf3 = std::numeric_limits<double>::infinity();
-      for(int rep = 0; rep < 3; rep++) {
-        double t = MLXWinogradTuner::scoreInputTransformForTesting(
-            MLXWinograd::InputTransform{}, 1, 19, 19, mi, true);
-        if(t < minOf3) minOf3 = t;
-      }
-
-      const double relErr = std::abs(parsedBaseline - minOf3) / minOf3;
-      testAssert(relErr < 0.25);
-      std::cout << "  baseline-consistency (gated) OK"
-                << " parsed=" << parsedBaseline
-                << " minOf3=" << minOf3
-                << " relErr=" << relErr << std::endl;
-
-      std::remove(tmpTunerFile.c_str());
-    }
-  }
-
-  {
-    // Per-shape numeric consistency — Test 2 from the shape-diagnostic spec.
-    // Asserts the dominant-shape median printed by flatSweepInput
-    // (shape_ms=c<C>:<ms>) is in the same ballpark as an independent
-    // reference measurement of the default InputTransform{} on that shape.
-    //
-    // IMPORTANT — cross-config comparison: parsedDominantMs is measured by
-    // flatSweepInput on the WINNER configuration (whatever the sweep
-    // selected). minOf3 is computed on the DEFAULT InputTransform{} via
-    // three independent scoreInputTransformPerShapeForTesting calls. These
-    // are not the same config, so the relative-error budget is necessarily
-    // loose. The budget covers:
-    //   - winner-vs-default speed gap (sweep can find configs 10-40%
-    //     faster than default on some shapes/hardware)
-    //   - selection bias on the min-of-3 reference (~5-10% low vs single)
-    //   - per-call noise floor (~10%)
-    // The 50% budget is intentionally conservative; this is a sanity-check
-    // that measurement is roughly working, not a tight precision check.
-    // Tighter precision checks belong in same-config stability tests.
-    //
-    // Coverage scope: input stage only. flatSweepOutput's per-shape fields
-    // are format-checked by the log-format test (gate
-    // KATAGO_MLX_WINOTUNER_RUN_LOG_FORMAT_TEST) but not consistency-
-    // checked here — symmetric output check is deferred.
-    //
-    // Gate is new (KATAGO_MLX_WINOTUNER_RUN_PER_SHAPE_TEST) and separate
-    // from the baseline-anchor gate above; this test runs an additional
-    // tuner sweep.
-    const char* gate = std::getenv("KATAGO_MLX_WINOTUNER_RUN_PER_SHAPE_TEST");
-    if(gate != nullptr && std::string(gate) == "1") {
-      MLXWinogradTuner::ModelInfoForTuning mi;
-      mi.trunkNumChannels    = 64;
-      mi.modelVersion        = 11;
-      // Synthetic single-shape histogram for the toy C=64 test model.
-      mi.conv3x3InputHistogram  = {{64, 1}};
-      mi.conv3x3OutputHistogram = {{64, 1}};
-
-      std::string tmpTunerFile = "/tmp/per_shape_consistency.txt";
-      std::remove(tmpTunerFile.c_str());
-
-      std::ostringstream captured;
-      Logger logger(nullptr, /*logToStdoutDefault=*/false,
-                    /*logToStderrDefault=*/false, /*logTimeDefault=*/false,
-                    /*logConfigContents=*/false);
-      logger.addOStream(captured);
-
-      (void)MLXWinogradTuner::loadOrAutoTune(
-          /*tunerFile=*/tmpTunerFile,
-          /*homeDataDirOverride=*/"",
-          /*gpuName=*/"AppleSilicon",
-          /*nnXLen=*/19, /*nnYLen=*/19, /*batchSize=*/1,
-          mi,
-          /*logger=*/&logger,
-          /*full=*/false,
-          /*reTune=*/true,
-          /*useFP16=*/true);
-
-      const std::string log = captured.str();
-      std::smatch m;
-      std::regex trunkRe(R"(flatSweepInput:[^\n]*shape_ms=c[0-9]+:([0-9]+\.[0-9]+))");
-      testAssert(std::regex_search(log, m, trunkRe));
-      const double parsedDominantMs = std::stod(m[1].str());
-
-      // Per-shape consistency: parse the dominant shape's median from
-      // the flatSweepInput log line (which used scoreInputTransformPerShape
-      // on the winner) and compare against scoreInputTransformPerShapeForTesting
-      // on the default InputTransform. Cross-config (winner vs default)
-      // so a wide relErr bound (<0.50) is appropriate.
-      std::vector<std::pair<int,double>> r1 =
-          MLXWinogradTuner::scoreInputTransformPerShapeForTesting(
-              MLXWinograd::InputTransform{}, 1, 19, 19, mi, true);
-      std::vector<std::pair<int,double>> r2 =
-          MLXWinogradTuner::scoreInputTransformPerShapeForTesting(
-              MLXWinograd::InputTransform{}, 1, 19, 19, mi, true);
-      std::vector<std::pair<int,double>> r3 =
-          MLXWinogradTuner::scoreInputTransformPerShapeForTesting(
-              MLXWinograd::InputTransform{}, 1, 19, 19, mi, true);
-      testAssert(!r1.empty() && !r2.empty() && !r3.empty());
-      // Each result has the same shapes in the same order; take the
-      // dominant (index 0) per-shape median across the 3 runs.
-      double minOf3 = std::min({r1[0].second, r2[0].second, r3[0].second});
-
-      const double relErr = std::abs(parsedDominantMs - minOf3) / minOf3;
-      // 50% budget — see comment block above for rationale on the loose
-      // bound (cross-config comparison + selection bias + noise).
-      testAssert(relErr < 0.50);
-      std::cout << "  per-shape dominant consistency (gated) OK"
-                << " parsed=" << parsedDominantMs
-                << " minOf3=" << minOf3
-                << " relErr=" << relErr << std::endl;
-
-      std::remove(tmpTunerFile.c_str());
-    }
-  }
-
-  {
-    // Per-shape scoring smoke test: verify that scoreInputTransformPerShape
-    // and scoreOutputUntransformPerShape return finite positive values for
-    // each planned shape with a default-constructed
-    // InputTransform/OutputUntransform on a tiny shape. Gated under the same
-    // env var as the other GPU-touching tests; ungated CI shouldn't pay for
-    // GPU work.
-    const char* gate = std::getenv("KATAGO_MLX_WINOTUNER_RUN_SWEEP_TEST");
-    if(gate != nullptr && std::string(gate) == "1") {
-      MLXWinogradTuner::ModelInfoForTuning mi;
-      mi.trunkNumChannels    = 64;
-      mi.modelVersion        = 11;
-      // Synthetic single-shape histogram for the toy C=64 test model.
-      mi.conv3x3InputHistogram  = {{64, 1}};
-      mi.conv3x3OutputHistogram = {{64, 1}};
-
-      std::vector<std::pair<int,double>> in =
-          MLXWinogradTuner::scoreInputTransformPerShapeForTesting(
-              MLXWinograd::InputTransform{}, 1, 19, 19, mi, true);
-      testAssert(!in.empty());
-      for(const auto& [c, v] : in) {
-        testAssert(c > 0);
-        testAssert(std::isfinite(v));
-        testAssert(v > 0.0);
-        testAssert(v < 1000.0);  // sanity: <1s per call on Apple Silicon
-      }
-
-      std::vector<std::pair<int,double>> out =
-          MLXWinogradTuner::scoreOutputUntransformPerShapeForTesting(
-              MLXWinograd::OutputUntransform{}, 1, 19, 19, mi, true);
-      testAssert(!out.empty());
-      for(const auto& [c, v] : out) {
-        testAssert(c > 0);
-        testAssert(std::isfinite(v));
-        testAssert(v > 0.0);
-        testAssert(v < 1000.0);
-      }
-      std::cout << "  per-shape scoring smoke (gated) OK"
-                << " in[0]=c" << in[0].first << ":" << in[0].second
-                << " out[0]=c" << out[0].first << ":" << out[0].second
-                << std::endl;
-    }
-  }
-
-  cout << "MLX Winograd tuner tests passed" << endl;
 }
 
 #endif // USE_MLX_BACKEND
